feat(server): migrate to postgresql

+++ server/app/src/main/kotlin/mu/muse/common/persistence/Page.kt
@@ -2,9 +2,9 @@ package mu.muse.common.persistence
 
 data class Page<T>(
     val content: List<T>,
-    val contentSize: Int,
-    val pageSize: Int,
-    val pageNumber: Int,
-    val totalElements: Int,
-    val totalPages: Int,
+    val contentSize: Long,
+    val pageSize: Long,
+    val pageNumber: Long,
+    val totalElements: Long,
+    val totalPages: Long,
 )
