+++ server/app/src/main/kotlin/mu/muse/usecase/access/instrument/InstrumentExtractor.kt
@@ -19,7 +19,7 @@ interface InstrumentExtractor {
 
     fun findByCriteria(criteria: Criteria): List<Instrument>
 
-    fun totalElements(): Int
+    fun totalElements(): Long
 
     data class Criteria(
         val name: InstrumentName?,
